updating of helper files to match templates

+++ measures/nrc_report_pricing/resources/NRCMeasureHelper.rb
@@ -2,55 +2,68 @@
 # only adds functionality where required.
 
 require_relative 'BTAPMeasureHelper'
-require 'fileutils'
 
-module NRCReportingMeasureHelper
+module NRCMeasureHelper
   include BTAPMeasureHelper
 end
 
-module NRCReportingMeasureTestHelper
+module NRCMeasureTestHelper
   include BTAPMeasureTestHelper
 
   # Define the output path. Set defaults and remove any existing outputs.
   @output_root_path = File.expand_path("#{File.expand_path(__dir__)}/../tests/output")
-  Dir.mkdir @output_root_path unless Dir.exists?(@output_root_path) 
+  Dir.mkdir @output_root_path unless Dir.exists?(@output_root_path)
   @output_path = @output_root_path
-  existing_folders = Dir.entries(@output_root_path) - ['.','..'] # Remove current folder above from list before deleting!
-  existing_folders.each do |entry|
-  folder_to_remove = File.expand_path("#{@output_root_path}/#{entry}")
-  puts "Removing existing output folder: #{folder_to_remove}".yellow
-  FileUtils.rm_rf(folder_to_remove)
+
+  # Remove the existing test results. Need to control when this is done as multiple test scripts could be
+  #  accessing the same path.
+  # Must call this in the test script.
+  def self.removeOldOutputs(before: Time.now)
+    existing_folders = Dir.entries(@output_path) - ['.', '..'] # Remove current folder above from list before deleting!
+    existing_folders.each do |entry|
+      folder_to_remove = File.expand_path("#{@output_path}/#{entry}")
+      puts "Checking existing output folder: #{folder_to_remove}".green
+    if File.mtime(folder_to_remove) < before
+        puts "Removing existing output folder: #{folder_to_remove}".yellow
+        FileUtils.rm_rf(folder_to_remove)
+    else
+        puts "Skipping existing output folder: #{folder_to_remove}".light_blue
+    end
+    end
   end
-  
+
   # Define methods to manage output folders.
   def self.resetOutputFolder
     @output_path = @output_root_path
   end
+
   def self.appendOutputFolder(folder)
     # Append name and validate if specified by the user
     path = @output_root_path + "/" + folder
-  validateOutputFolder(path)
+    validateOutputFolder(path)
   end
+
   def self.validateOutputFolder(path)
     # This should not be the root_folder.
-  # Also check if it exists.
-  # By default use the test method name.
-  path = File.expand_path(path)
-  if path == @output_root_path 
-    # Append the calling method name and re-validate (need to jump back two methods)
-    path = @output_root_path + "/" + caller_locations(1,2)[1].label.split.last
-    validateOutputFolder(path)
-  elsif File.exist?(path)
-    # Create a numbered subfolder. First check if there is a numbered folder.
-    path = path.split(/--/).first
-    count = Dir.glob("#{path}*").count
-    path = path + "--#{count}"
-    validateOutputFolder(path)
-  else
-    @output_path = path
-  end
+    # Also check if it exists.
+    # By default use the test method name.
+    path = File.expand_path(path)
+    if path == @output_root_path
+      # Append the calling method name and re-validate (need to jump back two methods)
+      path = @output_root_path + "/" + caller_locations(1, 2)[1].label.split.last
+      validateOutputFolder(path)
+    elsif File.exist?(path)
+      # Create a numbered subfolder. First check if there is a numbered folder.
+      path = path.split(/--/).first
+      count = Dir.glob("#{path}*").count
+      path = path + "--#{count}"
+      validateOutputFolder(path)
+    else
+      @output_path = path
+    end
     @output_path.to_s
   end
+
   def self.outputFolder
     @output_path.to_s
   end
@@ -60,59 +73,59 @@ module NRCReportingMeasureTestHelper
   @measure_path = File.expand_path("#{File.expand_path(__dir__)}/../tests")
   @test_summary_mdfile = @measure_path + "/" + "README.md"
   FileUtils.rm_rf(@test_summary_mdfile)
-  
+
   # Create initial file with title.
   @testSummaryTitleRequired = true
+
   def self.testSummaryTitleRequired
     return @testSummaryTitleRequired
   end
-  
+
   def self.setTestSummaryTitle(required)
     @testSummaryTitleRequired = required
   end
-  
+
   def self.testSummaryMDfile
     @test_summary_mdfile.to_s
   end
-  
-  # Test count
+
+  # Test count.
   @testSummaryCount = 1
-  def self.incrementTestSummaryCount
+   def self.incrementTestSummaryCount
     @testSummaryCount += 1
   end
-  
+
   def self.testSummaryCount
     return @testSummaryCount
   end
 
-  #
-  # Custom way to run a reporting measure in the test. Overwrites run_measure definition in BTAPMeasureTestHelper.
+  # Custom way to run a measure in the test. Overwrites run_measure definition in BTAPMeasureHelper.
   def run_measure(input_arguments, model)
-  
+
     # Provide feedback as to what is being done to teh terminal.
     puts "Running measure".green
     puts "  with input arguments".green + " #{input_arguments}".light_blue
     puts "  on model with".green + " #{model.modelObjects.count}".light_blue + " objects".green
-    puts "  from method".green + " #{caller_locations(1,1)[0].label.split.last}".light_blue
+    puts "  from method".green + " #{caller_locations(1, 1)[0].label.split.last}".light_blue
 
     # Set the output folder. This should be unique (check done in validateOutputFolder). Create if does not exist.
-    output_folder = NRCReportingMeasureTestHelper.outputFolder
-    output_folder = NRCReportingMeasureTestHelper.validateOutputFolder(output_folder)
+    output_folder = NRCMeasureTestHelper.outputFolder
+    output_folder = NRCMeasureTestHelper.validateOutputFolder(output_folder)
     Dir.mkdir(output_folder) unless Dir.exists?(output_folder)
-  
+
     # This will create a instance of the measure you wish to test. It does this based on the test class name.
     measure = get_measure_object()
     measure.use_json_package = @use_json_package
     measure.use_string_double = @use_string_double
-  
-    # Return false if can't
+
+    # Return false if can't.
     return false if false == measure
-  
-    # Now get the arguments and create a runner
+
+    # Now get the arguments and create a runner.
     arguments = measure.arguments()
     argument_map = OpenStudio::Measure.convertOSArgumentVectorToMap(arguments)
     runner = OpenStudio::Measure::OSRunner.new(OpenStudio::WorkflowJSON.new)
-  
+
     # Set the arguements in the argument map use json or real arguments.
     if @use_json_package
       argument = arguments[0].clone
@@ -130,81 +143,16 @@ module NRCReportingMeasureTestHelper
         argument_map[key] = argument
       end
     end
-  
-    # Get the e+ output requests, this will be done automatically by OS App and PAT
-    idf_output_requests = measure.energyPlusOutputRequests(runner, argument_map)
 
-    # Mimic the process of running this measure in OS App or PAT.
-    # Create a new folder to work in (if it does not exist)
-    if !File.exist?(output_folder)
-      FileUtils.mkdir_p(output_folder)
-    end
-    assert(File.exist?(output_folder))
-
-    # Remove existing report.
-    report_path = "#{output_folder}/report.html"
-    if File.exist?(report_path)
-      FileUtils.rm(report_path)
-    end
-
-    # If the updated model exists remove it.
-    model_out_path = "#{output_folder}/model.osm"
-    if File.exist?(model_out_path)
-      FileUtils.rm(model_out_path)
-    end
-
-    # Convert output requests to OSM for testing, OS App and PAT will add these to the E+ Idf.
-    workspace = OpenStudio::Workspace.new('Draft'.to_StrictnessLevel, 'EnergyPlus'.to_IddFileType)
-    workspace.addObjects(idf_output_requests)
-    rt = OpenStudio::EnergyPlus::ReverseTranslator.new
-    request_model = rt.translateWorkspace(workspace)
-
-    # Add requested outputs to model.
-    model.addObjects(request_model.objects)
-    model.save(model_out_path, true)
-    
-    sql_path = "#{output_folder}/run/eplusout.sql"
-    if ENV['OPENSTUDIO_TEST_NO_CACHE_SQLFILE']
-      if File.exist?(sql_path)
-        FileUtils.rm_f(sql_path)
-      end
-    end
-    
-    osw_path = File.join(output_folder, 'in.osw')
-    osw_path = File.absolute_path(osw_path.to_s)
-
-    workflow = OpenStudio::WorkflowJSON.new
-    workflow.setSeedFile(File.absolute_path(model_out_path))
-    epw_path = (model.getWeatherFile.url).to_s
-    workflow.setWeatherFile(File.absolute_path(epw_path))
-    workflow.saveAs(osw_path)
-
-    cli_path = OpenStudio.getOpenStudioCLI
-    cmd = "\"#{cli_path}\" run -w \"#{osw_path}\""
-    puts "Running openstudio with command: ".green + "\n#{cmd}".light_blue
-    OpenstudioStandards.run_command(cmd)
-  
-    # Set up runner, this will happen automatically when measure is run in PAT or OpenStudio.
-    runner.setLastOpenStudioModelPath(model_out_path)
-    runner.setLastEpwFilePath(epw_path)
-    runner.setLastEnergyPlusSqlFilePath(sql_path)
-  
-    # Temporarily change directory to the run directory and run the measure
-    start_dir = Dir.pwd
-    begin
-      Dir.chdir(output_folder)
-      measure.run(runner, argument_map)
-    ensure
-      Dir.chdir(start_dir)
-    end
-  
+    # Run the measure.
+    measure.run(model, runner, argument_map)
+    runner.result
     # Reset the output path to the root folder.
-    NRCReportingMeasureTestHelper.resetOutputFolder
+    NRCMeasureTestHelper.resetOutputFolder
 
-    # Add summary of test to README file
+    # Add summary of test to README file.
     measure_name = measure.name.gsub("_", " ").upcase
     reportCase(measure_name, output_folder.split('/').last, input_arguments)
-
     return runner
   end
 
@@ -212,33 +160,42 @@ module NRCReportingMeasureTestHelper
   def reportCase(measure_name, test_name, input_arguments)
 
     # File name defined above. Open for appending.
-    out_file = File.new("#{NRCReportingMeasureTestHelper.testSummaryMDfile}", "a")
+    out_file = File.new("#{NRCMeasureTestHelper.testSummaryMDfile}", "a")
 
-    # Only add the page title once.
-    if NRCReportingMeasureTestHelper.testSummaryTitleRequired
+    # Only add the page title once
+    if NRCMeasureTestHelper.testSummaryTitleRequired
       title = "# Summary Of Test Cases for '#{measure_name}' Measure"
       out_file.puts("#{title}")
       out_file.puts(" ")
     end
-  
+
     # Current test name.
     test_name = test_name.gsub("_", " ")
-    out_file.puts("## #{NRCReportingMeasureTestHelper.testSummaryCount} - #{test_name}")
-  
+    out_file.puts("## #{NRCMeasureTestHelper.testSummaryCount} - #{test_name}")
+
     # Create a table describing the case tested. Table header first.
     out_file.puts("| Test Argument | Test Value |")
     out_file.puts("| ------------- | ---------- |")
 
     # Table contents.
     input_arguments.each do |key, value|
-      out_file.puts("| #{key} |#{value} |")
+
+      if key.include? "json_input"
+        # Value is a string, has to be parsed and converted into hash.
+        value1=JSON.parse(value)
+        value1.each do |key, value|
+          out_file.puts("| #{key} |#{value} |")
+        end
+      else
+        out_file.puts("| #{key} |#{value} |")
+      end
     end
     out_file.puts(" ")
     out_file.close
-  
+
     # Update logical and counters.
-    NRCReportingMeasureTestHelper.setTestSummaryTitle(false)
-    NRCReportingMeasureTestHelper.incrementTestSummaryCount
+    NRCMeasureTestHelper.setTestSummaryTitle(false)
+    NRCMeasureTestHelper.incrementTestSummaryCount
   end
 
   # Fancy way of getting the measure object automatically. Added check for NRC in measure name.
@@ -261,6 +218,17 @@ module NRCReportingMeasureTestHelper
     else
       return nrc_measure
     end
+  end
+
+  # Load a test model (code that is common in a lot of test scripts). Returns the model object.
+  def load_test_osm(full_osm_model_path)
+
+    # Load the supplied osm.
+    translator = OpenStudio::OSVersion::VersionTranslator.new
+    path = OpenStudio::Path.new(full_osm_model_path)
+    model = translator.loadModel(path)
+    assert((not model.empty?))
+    model = model.get
   end
 end
